python/socialweb/requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==3.2.12
django-filter==21.1
gunicorn==20.1.0
Pillow==9.0.1
pytz==2022.1
sorl-thumbnail==12.8.0
sqlparse==0.4.2
typing-extensions==4.1.1
psycopg2==2.7.1
whitenoise==6.0.0
